Added User Roles Functionality
If a user is not logged in they only have access to the login and dashboard. If a user login as a viewer it will not have access to services page. also doesn't have access to edit services button in event form page. If it is logged in as editor then it will have access to services page and crud services